Point.value int to byte

+++ SudokuSloverHendler/BetterMatrix/AuxiliaryAlgorithms.cs
@@ -48,18 +48,18 @@ namespace SudokuSloverHendler.BetterMatrix
         public bool IsOneNullInSquare(PosSquare pos_s)
         { return this.IsOneNullInRange(new PosPoint(pos_s.i * 3, pos_s.j * 3), new PosPoint(pos_s.i * 3 + 2, pos_s.j * 3 + 2)); }
 
-        public Set<int> GetPossValueInPosPoint(PosPoint pos_p)
-        { return new Set<int>(this.matrix[pos_p.i, pos_p.j].set); }
+        public Set<byte> GetPossValueInPosPoint(PosPoint pos_p)
+        { return new Set<byte>(this.matrix[pos_p.i, pos_p.j].set); }
         public bool IsNullInPosPoint(PosPoint pos_p)
         { return this.matrix[pos_p.i, pos_p.j].value == 0; }
-        public bool IsPossibleValueInPos(PosPoint pos_p, int value)
+        public bool IsPossibleValueInPos(PosPoint pos_p, byte value)
         { return this.matrix[pos_p.i, pos_p.j].set.Contains(value); }
         public bool IsOnePossibleValueInPosPoint(PosPoint pos_p)
         { return this.matrix[pos_p.i, pos_p.j].set.Count() == 1; }
         public int GetFirstValueSetInPosPoint(PosPoint pos_p)
         { return this.matrix[pos_p.i, pos_p.j].set[0]; }
 
-        private Arrange<PosPoint> GetPossPosPointsInRange(PosPoint pos1, PosPoint pos2, int value)
+        private Arrange<PosPoint> GetPossPosPointsInRange(PosPoint pos1, PosPoint pos2, byte value)
         {
             Arrange<PosPoint> arr = new Arrange<PosPoint>();
             for (int i = pos1.i; i <= pos2.i; i++)
@@ -78,14 +78,14 @@ namespace SudokuSloverHendler.BetterMatrix
             return arr;
         }
 
-        public Arrange<PosPoint> GetPossPosPointsInHorizontalLine(int index, int value)
+        public Arrange<PosPoint> GetPossPosPointsInHorizontalLine(int index, byte value)
         { return this.GetPossPosPointsInRange(new PosPoint(index, 0), new PosPoint(index, size - 1), value); }
-        public Arrange<PosPoint> GetPossPosPointsInVerticalLine(int index, int value)
+        public Arrange<PosPoint> GetPossPosPointsInVerticalLine(int index, byte value)
         { return this.GetPossPosPointsInRange(new PosPoint(0, index), new PosPoint(size - 1, index), value); }
-        public Arrange<PosPoint> GetPossPosPointsInSquare(PosSquare pos_s, int value)
+        public Arrange<PosPoint> GetPossPosPointsInSquare(PosSquare pos_s, byte value)
         { return this.GetPossPosPointsInRange(new PosPoint(pos_s.i * 3, pos_s.j * 3), new PosPoint(pos_s.i * 3 + 2, pos_s.j * 3 + 2), value); }
 
-        private int GetCountPossiblePosPointInRange(PosPoint pos1, PosPoint pos2, int value)
+        private int GetCountPossiblePosPointInRange(PosPoint pos1, PosPoint pos2, byte value)
         {
             int count = 0;
             for (int i = pos1.i; i <= pos2.i; i++)
@@ -101,11 +101,11 @@ namespace SudokuSloverHendler.BetterMatrix
             return count;
         }
 
-        public int GetCountPossiblePosPointInHorizontalLine(int index, int value)
+        public int GetCountPossiblePosPointInHorizontalLine(int index, byte value)
         { return this.GetCountPossiblePosPointInRange(new PosPoint(index, 0), new PosPoint(index, size - 1), value); }
-        public int GetCountPossiblePosPointInVerticalLine(int index, int value)
+        public int GetCountPossiblePosPointInVerticalLine(int index, byte value)
         { return this.GetCountPossiblePosPointInRange(new PosPoint(0, index), new PosPoint(size - 1, index), value); }
-        public int GetCountPossiblePosPointInSquare(PosSquare pos_s, int value)
+        public int GetCountPossiblePosPointInSquare(PosSquare pos_s, byte value)
         { return this.GetCountPossiblePosPointInRange(new PosPoint(pos_s.i * 3, pos_s.j * 3), new PosPoint(pos_s.i * 3 + 2, pos_s.j * 3 + 2), value); }
 
         private PosPoint GetOneNullPosPointInRange(PosPoint pos1, PosPoint pos2)
@@ -130,7 +130,7 @@ namespace SudokuSloverHendler.BetterMatrix
         public PosPoint GetPosPointNullSquare(PosSquare pos_s)
         { return this.GetOneNullPosPointInRange(new PosPoint(pos_s.i * 3, pos_s.j * 3), new PosPoint(pos_s.i * 3 + 2, pos_s.j * 3 + 2)); }
 
-        private PosPoint GetFirstPossiblePosPointInRange(PosPoint pos1, PosPoint pos2, int value)
+        private PosPoint GetFirstPossiblePosPointInRange(PosPoint pos1, PosPoint pos2, byte value)
         {
             for (int i = pos1.i; i <= pos2.i; i++)
             {
@@ -148,11 +148,11 @@ namespace SudokuSloverHendler.BetterMatrix
             return new PosPoint(0, 0);
         }
 
-        public PosPoint GetFirstPossiblePosPointInHorizontalLine(int index, int value)
+        public PosPoint GetFirstPossiblePosPointInHorizontalLine(int index, byte value)
         { return this.GetFirstPossiblePosPointInRange(new PosPoint(index, 0), new PosPoint(index, size - 1), value); }
-        public PosPoint GetFirstPossiblePosPointInVerticalLine(int index, int value)
+        public PosPoint GetFirstPossiblePosPointInVerticalLine(int index, byte value)
         { return this.GetFirstPossiblePosPointInRange(new PosPoint(0, index), new PosPoint(size - 1, index), value); }
-        public PosPoint GetFirstPossiblePosPointInSquare(PosSquare pos_s, int value)
+        public PosPoint GetFirstPossiblePosPointInSquare(PosSquare pos_s, byte value)
         { return this.GetFirstPossiblePosPointInRange(new PosPoint(pos_s.i * 3, pos_s.j * 3), new PosPoint(pos_s.i * 3 + 2, pos_s.j * 3 + 2), value); }
     }
